pyproject.toml
[tool.poetry]
name = "imagelocation"
version = "0.1.0"
description = ""
authors = ["andrewmehta <andmehta@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^11.0.0"
pillow-heif = "^0.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
